% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_get_iteration.R
\name{case_get_iteration}
\alias{case_get_iteration}
\title{Geta particular iteration of an imputation case}
\usage{
case_get_iteration(x, imputation_case, unit_id, iter)
}
\arguments{
\item{x}{A data frame containing the original dataset.}

\item{imputation_case}{A data frame containing the imputed values, including iteration and variable names.}

\item{unit_id}{A string specifying the column name used as an identifier.}

\item{iter}{An integer specifying which iteration to apply.}
}
\value{
A data frame with imputed values for the specified iteration.
}
\description{
This function imputes missing values in a dataset based on imputation values from another dataset.
It allows specifying the iteration (\code{iter}) and a dynamic identifier column (\code{business_col}).
}
