% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_save.R
\name{case_save}
\alias{case_save}
\title{Save imputation case as Parquet file}
\usage{
case_save(x, imputation_case, save_path, unit_id, optimized = FALSE)
}
\arguments{
\item{x}{A \code{data.frame}. The complete dataset on which missigness and imputation are going to be generated.}

\item{imputation_case}{A list containing two elements: \code{imputations} and \code{meta}.}

\item{save_path}{A string specifying the file path to save the Parquet file.}

\item{unit_id}{A string or numeric column name/index used to uniquely identify rows (e.g., subject ID or timestamp).}

\item{optimized}{A logical indicating if optimized version of specifying of \code{imputations} is desired.}
}
\value{
None. Saves the \code{imputations} data frame as a Parquet file.
}
\description{
This function extracts the \code{imputations} element from a \code{imputation_case} object
then saves it as a Parquet file at the specified location.
}
\examples{
library(arrow)
imputation_case <- list(imputations = data.frame(ID = c(1, 2, 3), value = c(NA, 5, 8)),
                        meta = NULL)
case_save(imputation_case$imputations, "imp_case.parq")

}
