% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_missing_indicators.R
\name{create_missing_indicators}
\alias{create_missing_indicators}
\title{Add Missingness Indicators to Original Data}
\usage{
create_missing_indicators(missingness_scenario)
}
\arguments{
\item{missingness_scenario}{A list containing three elements: \code{original_data}, \code{miss_iterations} and \code{meta}.}
}
\value{
A dataframe that includes \code{original_data}, \code{iter}, and missingness indicators for variables that
have a non-zero number of missing generations.
}
\description{
This function creates a new dataframe that includes all columns from \code{original_data}
and additional columns indicating missingness for each variable found in \code{missing_iterations}.
}
\examples{
missingness_scenario <- list(
            original_data = data.frame(ID = c(1, 2, 3, 4), var1 = c(7, 5, 9, 11), var2 = c(13, 21, 7, 15)),
            miss_iterations = data.frame(ID = c(1, 1, 2, 1, 2), iter = c(1, 3, 1, 2, 2),
                          variable = c('var1', 'var2', 'var1', 'var1', 'var1')),
            meta = list(unit_id = "business_number"))
create_missingness_df(original_data)

}
