% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_plot.R
\name{iteration_plot}
\alias{iteration_plot}
\title{Plot Missingness Scenario}
\usage{
iteration_plot(x, iter = NULL)
}
\arguments{
\item{x}{A dataframe that contains missingness.}

\item{iter}{An optional value that specifies for which iteration the missingness plot is desired.}
}
\value{
A plot displaying the missingness pattern.
}
\description{
This function visualizes the missingness pattern in a dataset using \code{md.pattern()} from \code{mice} package.
}
\examples{
library(mice)
df <- data.frame(a = c(1, NA, 3), b = c(4, 5, NA), iter = rep(1, 3))
plot_missingness(df)

}
