% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_create_from_case.R
\name{out_create_from_case}
\alias{out_create_from_case}
\title{Identify and Extract Changed Values Between Original and Imputed DataFrames}
\usage{
out_create_from_case(x, imputed, unit_id)
}
\arguments{
\item{x}{A data frame containing original values.}

\item{imputed}{A data frame containing imputed values with the same structure as original, plus an iteration column \code{iter}.}

\item{unit_id}{A character string specifying the column name representing unique identifiers.}
}
\value{
A data frame with columns "unit_id", "iter", "variable", and "value", containing only changed values, sorted by iter and unit_id.
}
\description{
This function compares two data frames (original and imputed) and extracts only records and values of what changed.
}
\examples{
x <- data.frame(unit_id = c(1,2,3,4), variable_a = c(10,20,30,40), variable_b = c(5,15,25,35))
imputed <- data.frame(unit_id = c(1,2,3,4,5), iter = c(1,1,1,1,1), variable_a = c(10,25,30,45,50), variable_b = c(5,15,20,35,40))
detect_changes(x, imputed, "unit_id")

}
