% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_generate_and_run_imputation_case.R
\name{print.imputation_case}
\alias{print.imputation_case}
\title{Print method for "imputations" objects}
\usage{
\method{print}{imputation_case}(x, ...)
}
\arguments{
\item{x}{An object of class "imputations".}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for "imputations" objects
}
