% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_print.R
\name{print.missingness_scenario}
\alias{print.missingness_scenario}
\title{Print method for "missingness_scenario" objects}
\usage{
\method{print}{missingness_scenario}(x, ...)
}
\arguments{
\item{x}{An object of class "missingness_scenario".}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for "missingness_scenario" objects
}
