% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_get_iteration.R,
%   R/scenario_get_simulated_dataset.R
\name{scenario_get_iteration}
\alias{scenario_get_iteration}
\title{Get scenario iteration using missingness scenario}
\usage{
scenario_get_iteration(missingness_scenario, iter)

scenario_get_iteration(missingness_scenario, iter)
}
\arguments{
\item{missingness_scenario}{A list containing three elements: \code{original_data}, \code{scenario_iteration} and \code{meta}.}

\item{iter}{A single value specifying which iteration to retain.}
}
\value{
A data frame with missing values for the selected iteration.

A data frame with missing values for the selected iteration.
}
\description{
This function produces a dataset with missing values based on a \code{missingness_scenario} object.

This function produces a dataset with missing values based on a \code{missingness_scenario} object.
}
\examples{
missingness_scenario <- list(
  meta = NULL,
  miss_iterations = data.frame(ID = c(1, 2, 1, 2), iter = c(1, 1, 2, 2), value = c(NA, 5, 8, 5))
)

scenario_get_iteration(missingness_scenario, 3)

missingness_scenario <- list(
  original_data = data.frame(ID = c(1, 2, 3, 4), var1 = c(7, 5, 9, 11), var2 = c(13, 21, 7, 15)),
  miss_iterations = data.frame(ID = c(1, 1, 2, 1, 2), iter = c(1, 1, 1, 2, 2),
                               variable = c('var1', 'var2', 'var1', 'var1', 'var1')),
  meta = NULL
)

scenario_get_iteration(missingness_scenario, 3)

}
