% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_save.R
\name{scenario_save}
\alias{scenario_save}
\title{Save missingness_scenario as Parquet file}
\usage{
scenario_save(missingness_scenario, save_path, optimized = FALSE)
}
\arguments{
\item{missingness_scenario}{A list containing two elements: \code{meta} and \code{miss_iterations}.}

\item{save_path}{A string specifying the file path to save the Parquet file.}

\item{optimized}{A logical indicating if optimized version of specifying of \code{miss_iterations} is desired.}
}
\value{
None. Saves the \code{miss_iterations} data frame as a Parquet file.
}
\description{
This function extracts the \code{miss_iterations} element from a \code{missingness_scenario} object
then saves it as a Parquet file at the specified location.
}
\examples{
library(arrow)
missingness_scenario <- list(meta = NULL,
                             miss_iterations = data.frame(ID = c(1, 2, 3), value = c(NA, 5, 8)))
scenario_save(missingness_scenario, "miss_scenario.parquet")

}
